# lib/api-zod/src/generated/types/clientUser.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * ProcureTrack API - Purchase Order Management System
 * OpenAPI spec version: 0.1.0
 */

export interface ClientUser {
  id: string;
  email: string;
  /** @nullable */
  name?: string | null;
}