# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * ProcureTrack API - Purchase Order Management System
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummarySuppliersByStatus } from './dashboardSummarySuppliersByStatus';
import type { PurchaseOrder } from './purchaseOrder';

export interface DashboardSummary {
  totalOrders: number;
  activeOrders: number;
  completedOrders: number;
  draftOrders: number;
  cancelledOrders: number;
  totalSuppliers: number;
  suppliersByStatus: DashboardSummarySuppliersByStatus;
  unreadNotifications: number;
  recentOrders: PurchaseOrder[];
}