# lib/api-zod/src/generated/types/purchaseOrder.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * ProcureTrack API - Purchase Order Management System
 * OpenAPI spec version: 0.1.0
 */
import type { PurchaseOrderStatus } from './purchaseOrderStatus';

export interface PurchaseOrder {
  id: string;
  poNumber: string;
  title: string;
  /** @nullable */
  description?: string | null;
  status: PurchaseOrderStatus;
  clientId: string;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  clientEmail?: string | null;
  /** @nullable */
  totalAmount?: number | null;
  currency: string;
  supplierCount: number;
  createdAt: string;
  updatedAt: string;
}