# lib/api-zod/src/generated/types/purchaseOrderInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * ProcureTrack API - Purchase Order Management System
 * OpenAPI spec version: 0.1.0
 */
import type { PurchaseOrderInputStatus } from './purchaseOrderInputStatus';

export interface PurchaseOrderInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  clientId: string;
  totalAmount?: number;
  currency?: string;
  status?: PurchaseOrderInputStatus;
}