# lib/api-zod/src/generated/types/supplierUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * ProcureTrack API - Purchase Order Management System
 * OpenAPI spec version: 0.1.0
 */
import type { SupplierUpdateStatus } from './supplierUpdateStatus';

export interface SupplierUpdate {
  name?: string;
  trackingNumber?: string;
  trackingUrl?: string;
  status?: SupplierUpdateStatus;
  notes?: string;
  estimatedDelivery?: string;
}